+++ docs/conf.py
@@ -12,7 +12,7 @@ sys.path.insert(0, os.path.abspath('..'))
 
 project = 'agate'
 copyright = '2017, Christopher Groskopf'
-version = '1.9.1'
+version = '1.10.0'
 release = version
 
 # -- General configuration ---------------------------------------------------