+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 1.10.0 - April 27. 2024
 -----------------------
 
+-  feat: :meth:`.Table.from_csv` reads the file line by line. If ``sniff_limit=None``, it reads the file into memory once, instead of twice. If ``column_types`` is a :class:`.TypeTester`, it reads the file into memory. (#778)
 -  fix: Fix :meth:`.TableSet.print_structure` for nested tablesets. (#765)
 
    .. code-block:: python